requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
streamlit>=1.28.0
pyyaml>=6.0
tqdm>=4.66.0
